# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hKask API",
    "description": "A Minimal Viable Container for Agents - HTTP API",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.27.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "hKask API server"
    }
  ],
  "paths": {
...
    "/api/v1/curator/escalations": {
      "get": {
        "tags": [
          "curator"
        ],
        "operationId": "list_escalations",
        "responses": {
          "200": {
            "description": "List of pending escalations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListEscalationsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/curator/escalations/{id}/dismiss": {
      "post": {
        "tags": [
          "curator"
        ],
        "operationId": "dismiss_escalation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DismissEscalationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Escalation dismissed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DismissEscalationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Escalation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/curator/escalations/{id}/resolve": {
      "post": {
        "tags": [
          "curator"
        ],
        "operationId": "resolve_escalation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveEscalationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Escalation resolved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveEscalationResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Escalation not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/curator/metacognition": {
      "get": {
        "tags": [
          "curator"
        ],
        "operationId": "metacognition_status",
        "responses": {
          "200": {
            "description": "Curator metacognition status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MetacognitionStatusResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BotStatusReportResponse": {
        "type": "object",
        "required": [
          "bot_name",
          "status",
          "issues"
        ],
        "properties": {
          "bot_name": {
            "type": "string"
          },
          "issues": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "last_report": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "DismissEscalationRequest": {
        "type": "object",
        "required": [
          "dismissed_by"
        ],
        "properties": {
          "dismissed_by": {
            "type": "string"
          }
        }
      },
      "DismissEscalationResponse": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
      "EscalationEntryResponse": {
        "type": "object",
        "required": [
          "id",
          "template_id",
          "bot_id",
          "output",
          "confidence",
          "retry_count",
          "error_context",
          "created_at",
          "status"
        ],
        "properties": {
          "bot_id": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          },
          "created_at": {
            "type": "string"
          },
          "error_context": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "output": {
            "type": "string"
          },
          "resolved_at": {
            "type": [
              "string",
              "null"
            ]
          },
          "resolved_by": {
            "type": [
              "string",
              "null"
            ]
          },
          "retry_count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "status": {
            "type": "string"
          },
          "template_id": {
            "type": "string"
          }
        }
      },
      "EscalationStatsResponse": {
        "type": "object",
        "required": [
          "total",
          "pending",
          "resolved",
          "dismissed"
        ],
        "properties": {
          "dismissed": {
            "type": "integer",
            "format": "int64"
          },
          "pending": {
            "type": "integer",
            "format": "int64"
          },
          "resolved": {
            "type": "integer",
            "format": "int64"
          },
          "total": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "ListEscalationsResponse": {
        "type": "object",
        "required": [
          "escalations"
        ],
        "properties": {
          "escalations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EscalationEntryResponse"
            }
          }
        }
      },
...
      "MetacognitionStatusResponse": {
        "type": "object",
        "required": [
          "escalation_stats",
          "bot_reports"
        ],
        "properties": {
          "bot_reports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BotStatusReportResponse"
            }
          },
          "escalation_stats": {
            "$ref": "#/components/schemas/EscalationStatsResponse"
          }
        }
      },
...
      "ResolveEscalationRequest": {
        "type": "object",
        "required": [
          "resolved_by"
        ],
        "properties": {
          "resolved_by": {
            "type": "string"
          }
        }
      },
      "ResolveEscalationResponse": {
        "type": "object",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}